Moved subscription tests to a separate file
